+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fix [#195](https://github.com/shinolab/autd3/issues/195): missing setter for speed of sound in Geometry for C++/C#/Python
 - Fix [#196](https://github.com/shinolab/autd3/issues/196): modulation::Fourier must have one component at least
 - Fix [#197](https://github.com/shinolab/autd3/issues/197): Add ability to update Geometry in Simulator
+- Fix [#198](https://github.com/shinolab/autd3/issues/198): [C#] Make Controller::group and Gain::Group can use any type as key
 
 # 15.2.1