+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 15.3.1
+
+- Fix [#199](https://github.com/shinolab/autd3/issues/199): Add Device::translate_to and Device::rotate_to
+
 # 15.3.0
 
 - Raise minimum supported Python version to 3.10